Add a utility to create a normal or colorized depth image
