+++ source/Utilities/CMakeLists.txt
@@ -63,6 +63,7 @@ add_subdirectory(ChangeResolution)
 add_subdirectory(ChangeTransmitDelay)
 add_subdirectory(ColorImageUtility)
 add_subdirectory(DeviceInfoUtility)
+add_subdirectory(DepthImageUtility)
 add_subdirectory(ExternalCalUtility)
 add_subdirectory(FlashUtility)
 add_subdirectory(ImageCalUtility)
