+++ source/Utilities/shared/Io.hh
@@ -1,0 +1,121 @@
+/**
+ * @file shared/Io.hh
+ *
+ * Copyright 2013-2023
+ * Carnegie Robotics, LLC
+ * 4501 Hatfield Street, Pittsburgh, PA 15201
+ * https://www.carnegierobotics.com
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *     * Redistributions of source code must retain the above copyright
+ *       notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *     * Neither the name of the Carnegie Robotics, LLC nor the
+ *       names of its contributors may be used to endorse or promote products
+ *       derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL CARNEGIE ROBOTICS, LLC BE LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ **/
+
+#ifndef IO_HH
+#define IO_HH
+
+#include <string>
+
+namespace io
+{
+
+bool savePpm(const std::string& fileName,
+        uint32_t           width,
+        uint32_t           height,
+        const void* dataP)
+{
+    std::ofstream outputStream(fileName.c_str(), std::ios::out | std::ios::binary);
+
+    if (false == outputStream.good()) {
+        std::cerr << "Failed to open \"" << fileName << "\"" << std::endl;
+        return false;
+    }
+
+    const uint32_t imageSize = height * width * 3;
+
+    outputStream << "P6\n"
+        << width << " " << height << "\n"
+        << 0xFF << "\n";
+
+    outputStream.write(reinterpret_cast<const char*>(dataP), imageSize);
+
+    outputStream.close();
+    return true;
+}
+
+bool savePgm(const std::string& fileName,
+             uint32_t           width,
+             uint32_t           height,
+             uint32_t           bitsPerPixel,
+             const std::string& comment,
+             const void         *dataP)
+{
+    std::ofstream outputStream(fileName.c_str(), std::ios::binary | std::ios::out);
+
+    if (false == outputStream.good()) {
+		std::cerr << "Failed to open \"" << fileName << "\"" << std::endl;
+        return false;
+    }
+
+    const uint32_t imageSize = height * width;
+
+    switch(bitsPerPixel) {
+    case 8:
+    {
+
+        outputStream << "P5\n"
+                     << "#" << comment << "\n"
+                     << width << " " << height << "\n"
+                     << 0xFF << "\n";
+
+        outputStream.write(reinterpret_cast<const char*>(dataP), imageSize);
+
+        break;
+    }
+    case 16:
+    {
+        outputStream << "P5\n"
+                     << "#" << comment << "\n"
+                     << width << " " << height << "\n"
+                     << 0xFFFF << "\n";
+
+        const uint16_t *imageP = reinterpret_cast<const uint16_t*>(dataP);
+
+        for (uint32_t i=0; i<imageSize; ++i) {
+            uint16_t o = htons(imageP[i]);
+            outputStream.write(reinterpret_cast<const char*>(&o), sizeof(uint16_t));
+        }
+
+        break;
+    }
+    }
+
+    outputStream.close();
+    return true;
+}
+
+}
+
+#endif //IO_HH
+
